Former-commit-id: d692ea43bba683d0a49158001f9b4d7af8da72c1

+++ BE/spring/authservice/src/main/java/com/booklet/authservice/entity/Comment.java
@@ -12,7 +12,7 @@ public class Comment extends BaseTimeEntity {
     @Column(name = "comment_id")
     private Long commentId;
     @Column(nullable = false)
-    private int commentGroup; // 댓글이 속한 그룹 ( 모댓글의 commentsID )
+    private Long commentGroup; // 댓글이 속한 그룹 ( 모댓글의 commentsID )
     @Column(nullable = false)
     private int commentDepth; // 0 : 모댓글, 1 : 자식
     @Column(nullable = false)
